Fix type mismatch: change logIngestSummary parameters from int to long

+++ apps/reporter-service/src/main/java/com/novareport/reporter_service/service/RssIngestService.java
@@ -59,7 +59,7 @@ public class RssIngestService {
         }
 
         Map<String, NewsItem> deduped = deduplicateByHash(items);
-        int stored = persistNewItems(deduped);
+        long stored = persistNewItems(deduped);
         logIngestSummary(attempted, stored);
 
         return new IngestResult(attempted, stored);
@@ -105,7 +105,7 @@ public class RssIngestService {
         return toPersist.isEmpty() ? 0 : newsItemRepository.saveAll(toPersist).size();
     }
 
-    private void logIngestSummary(int attempted, int stored) {
+    private void logIngestSummary(long attempted, long stored) {
         double storageRatio = attempted == 0 ? 0 : (double) stored / attempted;
         double dedupeRatio = 1 - storageRatio;
         log.info(
